+++ src/ui/modals/keybinds.rs
@@ -10,6 +10,7 @@ use ratatui::{
     text::{Line, Text},
     widgets::{Block, Borders, Cell, Clear, Row, Table, TableState},
 };
+use strum::{IntoDiscriminant, VariantArray};
 
 use super::{Modal, RectExt};
 use crate::{
@@ -31,12 +32,23 @@ pub struct KeybindsModal {
 }
 
 trait KeybindsExt {
-    fn to_str(&self) -> impl Iterator<Item = (String, String, Cow<'static, str>)>;
+    fn sort_by_action(&self) -> impl Iterator<Item = (String, String, Cow<'static, str>)>;
 }
 
-impl<K: Display, V: Display + ToDescription> KeybindsExt for HashMap<K, V> {
-    fn to_str(&self) -> impl Iterator<Item = (String, String, Cow<'static, str>)> {
-        self.iter().map(|(key, value)| (key.to_string(), value.to_string(), value.to_description()))
+impl<K, ActionEnum, Discriminant> KeybindsExt for HashMap<K, ActionEnum>
+where
+    K: Display + std::cmp::PartialEq<K>,
+    ActionEnum: Display + ToDescription + IntoDiscriminant<Discriminant = Discriminant>,
+    Discriminant: VariantArray + std::cmp::PartialEq<Discriminant>,
+{
+    fn sort_by_action(&self) -> impl Iterator<Item = (String, String, Cow<'static, str>)> {
+        Discriminant::VARIANTS.iter().flat_map(|variant| {
+            self.iter()
+                .filter(|(_, v)| &v.discriminant() == variant)
+                .map(|(k, v)| (k.to_string(), v))
+                .sorted_by(|a, b| a.0.cmp(&b.0))
+                .map(|(k, v)| (k, v.to_string(), v.to_description()))
+        })
     }
 }
 
@@ -123,39 +135,15 @@ impl Modal for KeybindsModal {
         let keybinds = &app.config.keybinds;
         let header_style = app.config.theme.current_item_style;
 
-        let mut global: Vec<_> = keybinds.global.to_str().collect();
-        let mut navigation: Vec<_> = keybinds.navigation.to_str().collect();
-        let mut albums: Vec<_> = keybinds.albums.to_str().collect();
-        let mut artists: Vec<_> = keybinds.artists.to_str().collect();
-        let mut directories: Vec<_> = keybinds.directories.to_str().collect();
-        let mut playlists: Vec<_> = keybinds.playlists.to_str().collect();
-        let mut search: Vec<_> = keybinds.search.to_str().collect();
-        let mut queue: Vec<_> = keybinds.queue.to_str().collect();
-
-        global.sort_by_key(|(_, action, _)| action.to_lowercase());
-        navigation.sort_by_key(|(_, action, _)| action.to_lowercase());
-        albums.sort_by_key(|(_, action, _)| action.to_lowercase());
-        artists.sort_by_key(|(_, action, _)| action.to_lowercase());
-        directories.sort_by_key(|(_, action, _)| action.to_lowercase());
-        playlists.sort_by_key(|(_, action, _)| action.to_lowercase());
-        search.sort_by_key(|(_, action, _)| action.to_lowercase());
-        queue.sort_by_key(|(_, action, _)| action.to_lowercase());
+        let global = keybinds.global.sort_by_action().collect_vec();
+        let navigation = keybinds.navigation.sort_by_action().collect_vec();
+        let queue = keybinds.queue.sort_by_action().collect_vec();
 
         let rows = row_header(&global, "Global", header_style)
             .into_iter()
             .chain(row(&global, key_area.width, action_area.width, desc_area.width))
             .chain(row_header(&navigation, "Navigation", header_style))
             .chain(row(&navigation, key_area.width, action_area.width, desc_area.width))
-            .chain(row_header(&albums, "Albums", header_style))
-            .chain(row(&albums, key_area.width, action_area.width, desc_area.width))
-            .chain(row_header(&artists, "Artists", header_style))
-            .chain(row(&artists, key_area.width, action_area.width, desc_area.width))
-            .chain(row_header(&directories, "Directories", header_style))
-            .chain(row(&directories, key_area.width, action_area.width, desc_area.width))
-            .chain(row_header(&playlists, "Playlists", header_style))
-            .chain(row(&playlists, key_area.width, action_area.width, desc_area.width))
-            .chain(row_header(&search, "Search", header_style))
-            .chain(row(&search, key_area.width, action_area.width, desc_area.width))
             .chain(row_header(&queue, "Queue", header_style))
             .chain(row(&queue, key_area.width, action_area.width, desc_area.width))
             .collect_vec();